Introducing the LaTeX output with sympy library and the Taylor Polynomial with the Taylor expansion
In this commit in the polynomi.ipynb file the output in LaTeX has been added thanks to sympy. The polynomial divsion has it as default output, for the others operations, if you don't want to use print(), use the .latex() method.
A file named Taylor Polynomial.ipynb has been added to. Thanks to Plotly.py and another file created (Data Analysis 1.ipynb) it expands a function around a particoular point, creatin the Taylor
series expansion